add GetPath
